requirements.txt
flask>=2.3.3
openai==1.6.0
python-dotenv==1.0.0
werkzeug>=2.3.7